Wire up initialization logic
