+++ packages/common/src/index.ts
@@ -4,6 +4,7 @@ export * from './context'
 export * from './doctype'
 export * from './utils/doctype-utils'
 export * from './logger-provider'
+export * from './logger-provider-old' // TODO remove this
 export * from './pinning'
 export * from './doc-cache'
